+++ sleepyjs/src/parse.ts
@@ -1,25 +1,33 @@
 import { isArray, isString } from '@theroyalwhee0/istype';
 import { BlankCommand, Command, CommentCommand, NoopCommand, UserCommand } from './commands';
 import { ParseError } from './errors/parseerror';
-import { AwatedIterable, iterateLines } from './utilities/iter';
+import { bufferIterable, AwatedIterable, iterateLines } from './utilities/iter';
+
+export type ParseOptions = {
+    validateNow?: boolean // true = validate now, false or undefined validates as it streams
+};
 
 export interface Parsed {
-    rows: () => AsyncIterable<Command>,
+    rows: AsyncIterable<Command>,
 }
 
-export function parseText(input: string): Parsed {
-    return parseIterable(iterateLines(input));
+export async function parseText(input: string, options?: ParseOptions): Promise<Parsed> {
+    const iter = iterateLines(input);
+    return parseIterable(iter, options);
 }
 
-export function parseIterable(input: AwatedIterable<string>): Parsed {
+export async function parseIterable(input: AwatedIterable<string>, options?: ParseOptions): Promise<Parsed> {
+    async function* generator(): AsyncIterable<Command> {
+        let idx = 0;
+        for await (const item of input) {
+            yield parseSingleCommand(item, idx + 1);
+            idx++;
+        }
+    }
+    const iterable = generator();
+    const rows = options?.validateNow === true ? await bufferIterable(iterable) : iterable;
     return {
-        async *rows(): AsyncIterable<Command> {
-            let idx = 0;
-            for await (const item of input) {
-                yield parseSingleCommand(item, idx + 1);
-                idx++;
-            }
-        },
+        rows,
     };
 }