+++ sleepyjs/src/serialize.ts
@@ -6,7 +6,7 @@ export type SerializeOptions = {
 
 export async function* serializeCompiled(compiled: Compiled, options?: SerializeOptions): AsyncGenerator<string> {
     const pretty = options?.pretty ?? false;
-    for await (const value of compiled.rows()) {
+    for await (const value of compiled.rows) {
         if (pretty) {
             yield JSON.stringify(value, null, ' ').replace(/\s*\n\s*/g, ' ');
         } else {