+++ sleepyjs/test/compile.test.ts
@@ -1,10 +1,12 @@
-import { expect } from 'chai';
+import chai, { expect } from 'chai';
+import chaiAsPromised from 'chai-as-promised';
 import { describe, it } from 'mocha';
 import { BlankCommand, CommentCommand, NoopCommand, UserCommand } from '../src/commands';
 import { compileParsed } from '../src/compile';
 import { asyncIterabletoArray, mockParsed } from './mock';
 import { variesChaiEql } from './patch-eql';
 
+chai.use(chaiAsPromised);
 const varies = variesChaiEql();
 
 const sleepy_version = [0, 0, 1];
@@ -17,8 +19,12 @@ describe('compile', () => {
         it('should compile empty script', async () => {
             const parsed = mockParsed();
             const result = compileParsed(parsed);
-            expect(result).to.be.a('object');
-            const rows = await asyncIterabletoArray(result.rows());
+            expect(result).to.be.a('promise');
+            const compiled = await result;
+            expect(compiled).to.be.a('object');
+            expect(typeof compiled.rows).to.equal('object');
+            expect(compiled.rows[Symbol.asyncIterator]).to.be.a('function');
+            const rows = await asyncIterabletoArray(compiled.rows);
             expect(rows).to.eql([
                 ['@begin', sleepy_version],
                 ['@end'],
@@ -30,8 +36,8 @@ describe('compile', () => {
                 new CommentCommand(),
                 new NoopCommand(),
             ]);
-            const result = compileParsed(parsed, { optimize: true });
-            const rows = await asyncIterabletoArray(result.rows());
+            const compiled = await compileParsed(parsed, { optimize: true });
+            const rows = await asyncIterabletoArray(compiled.rows);
             expect(rows).to.eql([
                 ['@begin', sleepy_version],
                 ['@end'],
@@ -44,12 +50,12 @@ describe('compile', () => {
                 new NoopCommand(),
                 new NoopCommand(),
             ]);
-            const result = compileParsed(parsed, {
+            const compiled = await compileParsed(parsed, {
                 source: 'noop.zzz',
                 optimize: true,
                 info: true,
             });
-            const rows = await asyncIterabletoArray(result.rows());
+            const rows = await asyncIterabletoArray(compiled.rows);
             expect(rows).to.eql([
                 ['@begin', sleepy_version],
                 ['@meta', '@date', varies.to.match(/^\d+-\d+-\d+T\d+:\d+:\d+\.\d+Z$/)],
@@ -64,8 +70,8 @@ describe('compile', () => {
                 new NoopCommand(),
                 new NoopCommand(),
             ]);
-            const result = compileParsed(parsed, { optimize: false });
-            const rows = await asyncIterabletoArray(result.rows());
+            const compiled = await compileParsed(parsed, { optimize: false });
+            const rows = await asyncIterabletoArray(compiled.rows);
             expect(rows).to.eql([
                 ['@begin', sleepy_version],
                 ['@noop'],
@@ -81,8 +87,8 @@ describe('compile', () => {
                 new NoopCommand(),
                 new CommentCommand('# Hello World!'),
             ]);
-            const result = compileParsed(parsed, { optimize: false });
-            const rows = await asyncIterabletoArray(result.rows());
+            const compiled = await compileParsed(parsed, { optimize: false });
+            const rows = await asyncIterabletoArray(compiled.rows);
             expect(rows).to.eql([
                 ['@begin', sleepy_version],
                 ['@noop', '[]'],
@@ -99,8 +105,8 @@ describe('compile', () => {
                 new UserCommand('check', [{ 'gt': 32, 'lt': 293 }]),
                 new UserCommand('exit',),
             ]);
-            const result = compileParsed(parsed);
-            const rows = await asyncIterabletoArray(result.rows());
+            const compiled = await compileParsed(parsed);
+            const rows = await asyncIterabletoArray(compiled.rows);
             expect(rows).to.eql([
                 ['@begin', sleepy_version],
                 ['print', 'Hello World!'],
@@ -109,6 +115,26 @@ describe('compile', () => {
                 ['exit'],
                 ['@end'],
             ]);
+        });
+        it('should throw once iterated over', async () => {
+            const parsed = mockParsed([
+                new UserCommand('print', ['Hello World!']),
+                new Error('Something went wrong'),
+            ]);
+            const compiled = await compileParsed(parsed);
+            expect(compiled).to.be.a('object');
+            expect(typeof compiled.rows).to.equal('object');
+            expect(compiled.rows[Symbol.asyncIterator]).to.be.a('function');
+            const result = asyncIterabletoArray(compiled.rows);
+            expect(result).to.eventually.be.rejectedWith('Something went wrong');
+        });
+        it('should fail immediately when validateNow option set', async () => {
+            const parsed = mockParsed([
+                new UserCommand('print', ['Hello World!']),
+                new Error('Something went wrong'),
+            ]);
+            const result = compileParsed(parsed, { validateNow: true });
+            return expect(result).to.eventually.be.rejectedWith('Something went wrong');
         });
     });
 });